quic: cleanup and Bazel support (#90)

+++ src/net/aio/fd_aio.c
@@ -2,8 +2,7 @@
 
 /* generate one non-inline compiled copy of fd_aio_send */
 extern
-size_t
-fd_aio_send( fd_aio_t *        aio, 
+ulong
+fd_aio_send( fd_aio_t *        aio,
              fd_aio_buffer_t * batch,
-             size_t            batch_sz );
-
+             ulong             batch_sz );
